Changed dt from double to float. Separated Game Getters and Setters

+++ ZeldaClone/src/States/EditorState.h
@@ -7,7 +7,7 @@ class EditorState : public State
 public:
 	EditorState();
 
-	virtual void Update(const double& deltaTime) override;
+	virtual void Update(const float& deltaTime) override;
 	virtual void Render() override;
 
 	virtual bool OnEnter() override;
