+++ ZeldaClone/src/States/MenuState.h
@@ -18,7 +18,7 @@ public:
 	MenuState();
 	virtual ~MenuState() { Logger::Err(GetStateID() + " was Destroyed!"); };
 
-	virtual void Update(const double& deltaTime) override;
+	virtual void Update(const float& deltaTime) override;
 	virtual void Render() override;
 
 	virtual bool OnEnter() override;
