+++ ZeldaClone/src/States/SaveGameState.h
@@ -7,7 +7,7 @@ public:
 	SaveGameState();
 	~SaveGameState() {}
 
-	virtual void Update(const double& deltaTime) override;
+	virtual void Update(const float& deltaTime) override;
 	virtual void Render() override;
 
 	virtual bool OnEnter() override;
